+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [v0.4.7]
+### Fixed
+- Add stop and start tokens for `AnnotatedSpectrumDataset`, when available.
+- When `reverse` is used for the `PeptideTokenizer`, automatically reverse the decoded peptide.
 
 ## [v0.4.6]
 ### Added